Fix __slots__ crash
