+++ mypy/util.py
@@ -200,3 +200,17 @@ def correct_relative_import(cur_mod_id: str,
     if rel != 0:
         cur_mod_id = ".".join(parts[:-rel])
     return cur_mod_id + (("." + target) if target else ""), ok
+
+
+def replace_object_state(new: object, old: object) -> None:
+    # Copy state of old node to the new node. This varies depending on whether
+    # there is __slots__ and/or __dict__.
+    if hasattr(old, '__dict__'):
+        new.__dict__ = old.__dict__
+    if hasattr(old, '__slots__'):
+        # Use type.mro(...) since some classes override 'mro' with something different.
+        for base in type.mro(type(old)):
+            if '__slots__' in base.__dict__:
+                for attr in getattr(base, '__slots__'):
+                    if hasattr(old, attr):
+                        setattr(new, attr, getattr(old, attr))
